chore: Rename more limited 'core' to 'workspace'
GC is not solved yet
