+++ src/workspace/registry.rs
